requirements.txt
fvcore==0.1.5.post20210604
matplotlib==3.3.2
mmcv==1.3.5
numpy==1.18.5
opencv_python==4.5.2.54
Pillow==8.4.0
portalocker==2.0.0
requests==2.24.0
setproctitle==1.2.2
tifffile==2020.9.3
timm==0.3.2
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
